Server: Add protocol for parsing listfiles
This works by first parsing the file on disk, and using the resulting
parsed data to find the end of a command before a diff'd chunk.  That
way the parser can read an entire command in the changed chunk instead
of starting in the middle of an argument for example.

TODO: Instead of duplicating the parser, refactor the existing parser
into a visitor pattern or so to facilitate re-use in this context.